+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- added: (Ethereum) Add networkPrivacy user setting to enable privacy-enhanced fetch requests
+
 ## 4.70.0 (2026-01-16)
 
 - changed: (Polkadot) Update block explorer URL